split detour to 3 files
